feat d05ex02: added complete save()-method

+++ src_/day05/ex00/Chat/src/main/java/edu/school21/chat/models/Message.java
@@ -5,7 +5,7 @@ import java.time.format.DateTimeFormatter;
 
 public class Message {
 
-	private Integer id;
+	private Long id;
 
 	private String content;
 
@@ -15,7 +15,7 @@ public class Message {
 
 	private Chatroom chatroom;
 
-	public Message(Integer id, String content, User sender, Chatroom chatroom, LocalDateTime timestamp) {
+	public Message(Long id, String content, User sender, Chatroom chatroom, LocalDateTime timestamp) {
 		this.id = id;
 		this.content = content;
 		this.sender = sender;
@@ -23,11 +23,11 @@ public class Message {
         this.timestamp = timestamp;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
